refactor(backend): split invite persistence and models
