+++ backend/src/invites.rs
@@ -1,191 +1,23 @@
-use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine as _};
 use chrono::Utc;
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
-use sqlx::{Row, SqlitePool};
-use thiserror::Error;
+use sqlx::Row;
 
 mod handlers;
+mod model;
+mod store;
+
+use model::{clean_optional_text, hash_token, validate_redeemable_row};
 
 pub(crate) use handlers::{
     create_campaign_invite, inspect_campaign_invite, invite_error_code, invite_to_api_error,
     invite_token_from_next, list_campaign_invites, redeem_campaign_invite, revoke_campaign_invite,
 };
-
-#[derive(Debug, Clone)]
-pub(crate) struct InviteStore {
-    pool: SqlitePool,
-}
+pub(crate) use model::{
+    CampaignInvite, CreateInviteInput, InviteRole, InviteStatus, InviteStoreError, PublicInvite,
+    RedeemedInvite,
+};
+pub(crate) use store::InviteStore;
 
 impl InviteStore {
-    pub(crate) async fn initialize(pool: SqlitePool) -> Result<Self, InviteStoreError> {
-        let store = Self { pool };
-        let mut tx = store.pool.begin().await?;
-        sqlx::query(
-            r#"
-            CREATE TABLE IF NOT EXISTS campaign_invites (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                campaign_id INTEGER NOT NULL REFERENCES campaign_presets(id) ON DELETE CASCADE,
-                token_hash TEXT NOT NULL UNIQUE,
-                token_hint TEXT NOT NULL,
-                role TEXT NOT NULL CHECK (role = 'player'),
-                expires_at INTEGER,
-                max_uses INTEGER CHECK (max_uses IS NULL OR max_uses > 0),
-                use_count INTEGER NOT NULL DEFAULT 0 CHECK (use_count >= 0),
-                created_by_user_id INTEGER REFERENCES users(id) ON DELETE SET NULL,
-                revoked_at TEXT,
-                created_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-            )
-            "#,
-        )
-        .execute(&mut *tx)
-        .await?;
-        sqlx::query(
-            r#"
-            CREATE TABLE IF NOT EXISTS campaign_invite_redemptions (
-                invite_id INTEGER NOT NULL REFERENCES campaign_invites(id) ON DELETE CASCADE,
-                user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-                redeemed_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP,
-                PRIMARY KEY (invite_id, user_id)
-            )
-            "#,
-        )
-        .execute(&mut *tx)
-        .await?;
-        sqlx::query(
-            r#"
-            CREATE TABLE IF NOT EXISTS invite_restricted_users (
-                user_id INTEGER PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
-                created_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-            )
-            "#,
-        )
-        .execute(&mut *tx)
-        .await?;
-        sqlx::query(
-            "CREATE INDEX IF NOT EXISTS campaign_invites_campaign_id_idx ON campaign_invites(campaign_id)",
-        )
-        .execute(&mut *tx)
-        .await?;
-        tx.commit().await?;
-        Ok(store)
-    }
-
-    pub(crate) async fn create_invite(
-        &self,
-        campaign_id: i64,
-        created_by_user_id: i64,
-        input: &CreateInviteInput,
-        raw_token: &str,
-    ) -> Result<CampaignInvite, InviteStoreError> {
-        let token_hash = hash_token(raw_token);
-        let token_hint = raw_token.chars().take(8).collect::<String>();
-        let result = sqlx::query(
-            r#"
-            INSERT INTO campaign_invites (
-                campaign_id, token_hash, token_hint, role, expires_at,
-                max_uses, created_by_user_id
-            ) VALUES (?, ?, ?, 'player', ?, ?, ?)
-            "#,
-        )
-        .bind(campaign_id)
-        .bind(token_hash)
-        .bind(token_hint)
-        .bind(input.expires_at)
-        .bind(input.max_uses)
-        .bind(created_by_user_id)
-        .execute(&self.pool)
-        .await?;
-        self.find_invite_by_id(result.last_insert_rowid())
-            .await?
-            .ok_or(InviteStoreError::InviteNotFound)
-    }
-
-    pub(crate) async fn list_invites(
-        &self,
-        campaign_id: i64,
-    ) -> Result<Vec<CampaignInvite>, InviteStoreError> {
-        let rows = sqlx::query(
-            r#"
-            SELECT campaign_invites.id, campaign_invites.campaign_id,
-                   campaign_presets.display_name AS campaign_display_name,
-                   campaign_presets.room_slug, campaign_invites.token_hint,
-                   campaign_invites.role, campaign_invites.expires_at,
-                   campaign_invites.max_uses, campaign_invites.use_count,
-                   campaign_invites.revoked_at, campaign_invites.created_at
-            FROM campaign_invites
-            INNER JOIN campaign_presets ON campaign_presets.id = campaign_invites.campaign_id
-            WHERE campaign_invites.campaign_id = ?
-            ORDER BY campaign_invites.created_at DESC, campaign_invites.id DESC
-            "#,
-        )
-        .bind(campaign_id)
-        .fetch_all(&self.pool)
-        .await?;
-        rows.into_iter().map(invite_from_row).collect()
-    }
-
-    pub(crate) async fn revoke_invite(
-        &self,
-        campaign_id: i64,
-        invite_id: i64,
-    ) -> Result<(), InviteStoreError> {
-        let result = sqlx::query(
-            r#"
-            UPDATE campaign_invites
-            SET revoked_at = COALESCE(revoked_at, CURRENT_TIMESTAMP)
-            WHERE id = ? AND campaign_id = ?
-            "#,
-        )
-        .bind(invite_id)
-        .bind(campaign_id)
-        .execute(&self.pool)
-        .await?;
-        if result.rows_affected() == 0 {
-            return Err(InviteStoreError::InviteNotFound);
-        }
-        Ok(())
-    }
-
-    pub(crate) async fn inspect_invite(
-        &self,
-        raw_token: &str,
-    ) -> Result<PublicInvite, InviteStoreError> {
-        let row = sqlx::query(
-            r#"
-            SELECT campaign_invites.id, campaign_invites.campaign_id,
-                   campaign_presets.display_name AS campaign_display_name,
-                   campaign_presets.room_slug, campaign_presets.is_archived,
-                   campaign_invites.token_hint, campaign_invites.role,
-                   campaign_invites.expires_at, campaign_invites.max_uses,
-                   campaign_invites.use_count, campaign_invites.revoked_at,
-                   campaign_invites.created_at
-            FROM campaign_invites
-            INNER JOIN campaign_presets ON campaign_presets.id = campaign_invites.campaign_id
-            WHERE campaign_invites.token_hash = ?
-            "#,
-        )
-        .bind(hash_token(raw_token))
-        .fetch_optional(&self.pool)
-        .await?
-        .ok_or(InviteStoreError::InviteNotFound)?;
-        let is_archived = row.get::<i64, _>("is_archived") != 0;
-        let invite = invite_from_row(row)?;
-        let status = if is_archived {
-            InviteStatus::Archived
-        } else {
-            invite.status
-        };
-        Ok(PublicInvite {
-            campaign_id: invite.campaign_id,
-            campaign_display_name: invite.campaign_display_name,
-            room_slug: invite.room_slug,
-            role: invite.role,
-            expires_at: invite.expires_at,
-            status,
-        })
-    }
-
     pub(crate) async fn redeem_invite(
         &self,
         raw_token: &str,
@@ -343,189 +175,6 @@ impl InviteStore {
             room_slug,
         })
     }
-
-    pub(crate) async fn is_user_invite_restricted(
-        &self,
-        user_id: i64,
-    ) -> Result<bool, InviteStoreError> {
-        let count = sqlx::query_scalar::<_, i64>(
-            "SELECT COUNT(*) FROM invite_restricted_users WHERE user_id = ?",
-        )
-        .bind(user_id)
-        .fetch_one(&self.pool)
-        .await?;
-        Ok(count > 0)
-    }
-
-    async fn find_invite_by_id(
-        &self,
-        invite_id: i64,
-    ) -> Result<Option<CampaignInvite>, InviteStoreError> {
-        let row = sqlx::query(
-            r#"
-            SELECT campaign_invites.id, campaign_invites.campaign_id,
-                   campaign_presets.display_name AS campaign_display_name,
-                   campaign_presets.room_slug, campaign_invites.token_hint,
-                   campaign_invites.role, campaign_invites.expires_at,
-                   campaign_invites.max_uses, campaign_invites.use_count,
-                   campaign_invites.revoked_at, campaign_invites.created_at
-            FROM campaign_invites
-            INNER JOIN campaign_presets ON campaign_presets.id = campaign_invites.campaign_id
-            WHERE campaign_invites.id = ?
-            "#,
-        )
-        .bind(invite_id)
-        .fetch_optional(&self.pool)
-        .await?;
-        row.map(invite_from_row).transpose()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct CreateInviteInput {
-    pub(crate) expires_at: Option<i64>,
-    pub(crate) max_uses: Option<i64>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub(crate) enum InviteRole {
-    Player,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub(crate) enum InviteStatus {
-    Active,
-    Revoked,
-    Expired,
-    Exhausted,
-    Archived,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct CampaignInvite {
-    pub(crate) id: i64,
-    pub(crate) campaign_id: i64,
-    pub(crate) campaign_display_name: String,
-    pub(crate) room_slug: String,
-    pub(crate) token_hint: String,
-    pub(crate) role: InviteRole,
-    pub(crate) expires_at: Option<i64>,
-    pub(crate) max_uses: Option<i64>,
-    pub(crate) use_count: i64,
-    pub(crate) status: InviteStatus,
-    pub(crate) revoked_at: Option<String>,
-    pub(crate) created_at: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct PublicInvite {
-    pub(crate) campaign_id: i64,
-    pub(crate) campaign_display_name: String,
-    pub(crate) room_slug: String,
-    pub(crate) role: InviteRole,
-    pub(crate) expires_at: Option<i64>,
-    pub(crate) status: InviteStatus,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct RedeemedInvite {
-    pub(crate) user_id: i64,
-    pub(crate) campaign_id: i64,
-    pub(crate) campaign_display_name: String,
-    pub(crate) room_slug: String,
-}
-
-fn validate_redeemable_row(
-    row: &sqlx::sqlite::SqliteRow,
-    now: i64,
-) -> Result<(), InviteStoreError> {
-    if row.get::<i64, _>("is_archived") != 0 {
-        return Err(InviteStoreError::CampaignArchived);
-    }
-    if row.get::<Option<String>, _>("revoked_at").is_some() {
-        return Err(InviteStoreError::Revoked);
-    }
-    if row
-        .get::<Option<i64>, _>("expires_at")
-        .is_some_and(|expires_at| expires_at <= now)
-    {
-        return Err(InviteStoreError::Expired);
-    }
-    let max_uses = row.get::<Option<i64>, _>("max_uses");
-    let use_count = row.get::<i64, _>("use_count");
-    if max_uses.is_some_and(|max_uses| use_count >= max_uses) {
-        return Err(InviteStoreError::Exhausted);
-    }
-    Ok(())
-}
-
-fn invite_from_row(row: sqlx::sqlite::SqliteRow) -> Result<CampaignInvite, InviteStoreError> {
-    let role = match row.get::<String, _>("role").as_str() {
-        "player" => InviteRole::Player,
-        _ => return Err(InviteStoreError::InvalidIdentity),
-    };
-    let expires_at = row.get::<Option<i64>, _>("expires_at");
-    let max_uses = row.get::<Option<i64>, _>("max_uses");
-    let use_count = row.get::<i64, _>("use_count");
-    let revoked_at = row.get::<Option<String>, _>("revoked_at");
-    let status = if revoked_at.is_some() {
-        InviteStatus::Revoked
-    } else if expires_at.is_some_and(|expires_at| expires_at <= Utc::now().timestamp()) {
-        InviteStatus::Expired
-    } else if max_uses.is_some_and(|max_uses| use_count >= max_uses) {
-        InviteStatus::Exhausted
-    } else {
-        InviteStatus::Active
-    };
-    Ok(CampaignInvite {
-        id: row.get("id"),
-        campaign_id: row.get("campaign_id"),
-        campaign_display_name: row.get("campaign_display_name"),
-        room_slug: row.get("room_slug"),
-        token_hint: row.get("token_hint"),
-        role,
-        expires_at,
-        max_uses,
-        use_count,
-        status,
-        revoked_at,
-        created_at: row.get("created_at"),
-    })
-}
-
-fn clean_optional_text(value: Option<&str>) -> Option<String> {
-    value
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn hash_token(raw_token: &str) -> String {
-    URL_SAFE_NO_PAD.encode(Sha256::digest(raw_token.as_bytes()))
-}
-
-#[derive(Debug, Error)]
-pub(crate) enum InviteStoreError {
-    #[error("invite link is invalid")]
-    InviteNotFound,
-    #[error("invite link has been revoked")]
-    Revoked,
-    #[error("invite link has expired")]
-    Expired,
-    #[error("invite link has reached its maximum uses")]
-    Exhausted,
-    #[error("campaign is archived")]
-    CampaignArchived,
-    #[error("identity is already linked to another account")]
-    IdentityMismatch,
-    #[error("identity is invalid")]
-    InvalidIdentity,
-    #[error("invite cannot reactivate a privileged identity")]
-    PrivilegedIdentity,
-    #[error("database error: {0}")]
-    Sqlx(#[from] sqlx::Error),
 }
 
 #[cfg(test)]
